Drop down menu for AD
This is a working progress and I could use some help getting the buttons to do what they say continue, Exit etc... I am not sure how to tie the drop down option to the continue button I'm having some issues getting it to function. What is displayed is as far as I have gotten.